requirements.txt
gradio==5.44.1
langchain-core==0.3.74
loguru==0.7.3
langchain==0.3.27
langgraph==0.6.4
python-dotenv==1.1.1
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-google-genai==2.1.9
sentence-transformers==5.1.0
chromadb==1.0.20
rank_bm25==0.2.2